requirements.txt
streamlit
anthropic
spacy
langdetect
python-docx
pypdf
deep-translator
